pyproject.toml
[project]
name = "honeycomb-video-streamer"
dynamic = ["version"]

[tool.poetry]
name = "honeycomb-video-streamer"
version = "2.1.0"
authors = ["Paul DeCoursey <paul.decoursey@wildflowerschools.org>"]
maintainers = ["Benjamin Jaffe-Talberg <ben.talberg@wildflowerschools.org>"]
description = "Python tools to prepare videos for streaming from honeycomb"
keywords = ["video"]
repository = "https://github.com/WildflowerSchools/honeycomb-video-streamer"
readme = "README.md"
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python",
]

[tool.poetry.dependencies]
auth0-python = "^3.24.0"
boto3 = "^1.18.0"
click = ">=8.0"
ffmpeg-python = "^0.2.0"
numpy = "^1.21"
pandas = "^1.3.0"
python = "^3.8"
python-dotenv = ">=0.19.0"
toml = "^0.10.2"
wildflower-honeycomb-sdk = "^0.7.3"
wf-video-io = "^3.2.0"
sqlalchemy = "^1.4.45"
psycopg2 = "^2.9.5"
uvicorn = "^0.20.0"
wf-fastapi-auth0 = "^1.0.2"
python-jose = "^3.3.0"
pydantic = "^1.10.4"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
pylint = "^2.15.0"
pytest = "^7.2.0"

[tool.black]
line-length = 120
target-version = ['py38']

[tool.setuptools]
py-modules = []

[build-system]
requires = ["poetry-core", "setuptools"]
build-backend = "poetry.core.masonry.api"